+++ test/src/unit_random.cpp
@@ -25,7 +25,7 @@ TEST_CASE("unit/cdbn/random/mnist/1", "[cdbn][rectifier][svm][unit]") {
             , dll::conv_rbm_desc_square<20, 12, 20, 10, dll::parallel_mode, dll::momentum, dll::batch_size<10>>::layer_t
         >>::dbn_t;
 
-    auto dataset = mnist::read_dataset_direct<std::vector, etl::fast_dyn_matrix<double, 1, 28, 28>>(75);
+    auto dataset = mnist::read_dataset_direct<std::vector, etl::fast_dyn_matrix<float, 1, 28, 28>>(75);
     REQUIRE(!dataset.training_images.empty());
 
     mnist::binarize_dataset(dataset);
@@ -52,7 +52,7 @@ TEST_CASE("unit/cdbn/random/mnist/2", "[cdbn][rectifier][svm][unit]") {
             , dll::dyn_conv_rbm_desc<dll::parallel_mode, dll::momentum>::layer_t
         >>::dbn_t;
 
-    auto dataset = mnist::read_dataset_3d<std::vector, etl::dyn_matrix<double, 3>>(75);
+    auto dataset = mnist::read_dataset_3d<std::vector, etl::dyn_matrix<float, 3>>(75);
     REQUIRE(!dataset.training_images.empty());
 
     mnist::binarize_dataset(dataset);
